From develop to main.

+++ client/src/clientComponents/Card/Card.jsx
@@ -4,37 +4,58 @@ import { useEffect, useState } from "react";
 import { addItemsActions, deleteItemActions } from "../../redux/foodActions.js";
 import { useSelector, useDispatch } from "react-redux";
 import { useAuth0 } from "@auth0/auth0-react";
-import Swal from 'sweetalert2';
+import Swal from "sweetalert2";
 import axios from "axios";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faLeaf, faPlantWilt, faMillSign, faBreadSlice, faBowlFood, faPizzaSlice } from "@fortawesome/free-solid-svg-icons";
+
+import meat from "../../assets/categories/meat.png";
+import pastas from "../../assets/categories/pastas.png";
+import salad from "../../assets/categories/salad.png";
+import vegetarian from "../../assets/diets/vegetarian.png";
+import vegan from "../../assets/diets/vegan.png";
+import sinLactosa from "../../assets/diets/sinLactosa.png";
+import sinTacc from "../../assets/diets/sinTacc.png";
 
 
-export default function Card({ id, name, image, final_price, allItems }) {
+export default function Card({ id, name, image, initial_price, final_price, category, diets, discount, status, allItems }) {
   const [isItem, setIsItem] = useState(false);
-  const { isAuthenticated,user } = useAuth0();
-  const [quantity,setQuantity]=useState(1)
+  const { isAuthenticated, user } = useAuth0();
+  const [quantity, setQuantity] = useState(1);
+  
+  let categoryIcon;
+  if (category === "Carnes") categoryIcon = <img className={style.categoryIcon} src={meat}/>;
+  if (category === "Ensaladas") categoryIcon = <img className={style.categoryIcon} src={salad}/>;
+  if (category === "Pastas") categoryIcon = <img className={style.categoryIcon} src={pastas}/>;
 
-  console.log(allItems);
+  const dietsIcons = diets.map((diet, index) => {
+    if (diet === "Sin TACC") diet = <img key={index} className={style.dietsIcon} src={sinTacc}/>;
+    if (diet === "Vegetariano") diet = <img key={index} className={style.dietsIcon} src={vegetarian}/>;
+    if (diet === "Vegano") diet = <img key={index} className={style.dietsIcon} src={vegan}/>;
+    if (diet === "Sin Lactosa") diet = <img key={index} className={style.dietsIcon} src={sinLactosa}/>;
+    return diet;
+  }); 
+
+
   const dispatch = useDispatch();
   useEffect(() => {
     allItems.map((item) => {
       if (item.name == name) {
         setIsItem(true);
-        setQuantity(item.quantity)
+        setQuantity(item.quantity);
       }
     });
   }, []);
   const handleClick = (e) => {
-    // if (!isAuthenticated) {
-    //   //alert("¡Cuidado! Logueate antes de agregar productos a tu carrito de compras. ¡Gracias!");
-    //   Swal.fire(
-    //     '¡Cuidado!',
-    //     'Logueate antes de agregar productos a tu carrito de compras.',
-    //     'success'
-    //   )
-    // } else {
+    if (!isAuthenticated) {
+      Swal.fire(
+        "¡Cuidado!",
+        "Loguéate antes de agregar productos a tu carrito de compras.",
+        "success"
+      );
+    } else {
       if (isItem) {
         setIsItem(false), dispatch(deleteItemActions(id));
-        //-------------------------
         const bodyDeleteItem = {
           userEmail: user?.email,
           FoodId: id,
@@ -42,32 +63,38 @@ export default function Card({ id, name, image, final_price, allItems }) {
         axios
           .delete("/item", { data: bodyDeleteItem })
           .catch((error) => console.log(error));
-        //-------------------------
       } else {
         setIsItem(true);
         const amount = final_price * parseInt(quantity);
         dispatch(
           addItemsActions({ id, name, image, final_price, quantity, amount })
         );
-         //-------------------------
-         const bodyAddItem = {
+        const bodyAddItem = {
           userEmail: user?.email,
           FoodId: id,
           quantity,
           final_price,
         };
         axios.post("/item", bodyAddItem).catch((error) => console.log(error));
-        //-------------------------
       }
-    // }
+    }
   };
-  const updateQuantity=(e)=>{
-    const quantity=parseInt(e.target.value);
+
+  const updateQuantity = (e) => {
+    const quantity = parseInt(e.target.value);
     const amount = final_price * quantity;
-    setQuantity(quantity)
-    dispatch(deleteItemActions(id))
-    dispatch(addItemsActions({id,name,image,final_price,quantity:quantity,amount:amount}))
-    //-------------------------
+    setQuantity(quantity);
+    dispatch(deleteItemActions(id));
+    dispatch(
+      addItemsActions({
+        id,
+        name,
+        image,
+        final_price,
+        quantity: quantity,
+        amount: amount,
+      })
+    );
     const bodyUpdateItem = {
       userEmail: user?.email,
       FoodId: id,
@@ -75,32 +102,40 @@ export default function Card({ id, name, image, final_price, allItems }) {
       final_price,
     };
     axios.put("/item", bodyUpdateItem).catch((error) => console.log(error));
-    //-------------------------
-  }
+  };
+
+
+  /* RETURN */
   return (
     <div className={style.card}>
-      <NavLink to={`/detail/${id}`}>
+      <NavLink className={style.NavLink} to={`/detail/${id}`}>
         <div>
           <img src={image} alt="img not found" className={style.card_img} />
         </div>
         <div className={style.txt}>
-          <h2>{name}</h2>
+          <h2>{name} </h2>
+          <span>{categoryIcon} | {dietsIcons.map((dietIcon, index) => <span key={index}>{dietIcon}</span>)}</span>
         </div>
       </NavLink>
+
       <div className={style.p}>
         <p>${final_price}</p>
       </div>
       <div className={style.inputagregar}>
-        <button className={style.btncar} onClick={handleClick}>{isItem ? "Agregado" : "Agregar"}</button>
-        {isItem?<input className={style.detailinput} type="number" min='1' value={quantity} onChange={updateQuantity}/>:null}
+        <button className={style.btncar} onClick={handleClick}>
+          {isItem ? "Agregado" : "Agregar"}
+        </button>
+        {isItem ? (
+          <input
+            className={style.detailinput}
+            type="number"
+            min="1"
+            value={quantity}
+            onChange={updateQuantity}
+          />
+        ) : null}
       </div>
-      {/* <p>
-        Dietas:{" "}
-        {diets.map((diet) => (
-          <span>{diet}</span>
-        ))}
-      </p>
-      <p>Categoría: {category}</p> */}
+
     </div>
   );
 }
